backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
cryptography==41.0.3
python-dotenv==1.0.0
pytest==7.4.0
flask-sqlalchemy==3.0.5
flask-migrate==4.0.4
flask-jwt-extended==4.5.2
flask-bcrypt==1.0.1
flask-mail==0.9.1
psycopg2-binary==2.9.7
pyotp==2.9.0
qrcode==7.4.2
pillow==10.0.0